hps: don't retry every register read
The retry at this layer doesn't serve any purpose. It makes unit testing
more difficult and it makes it harder to reason about the retry and
delay behaviour at the layer above, in hps_impl.cc.

The reason for the retry noted in the comment (bank ready on MCP2221) is
not particularly important, since we generally don't use the C++ host
tools with HPS proto2 anymore. Plus, the existing retry logic in
WaitForBankReady() correctly handles errors from the MCP2221 while HPS
is NACK'ing. So there doesn't appear to be any reason to keep the retry.

BUG=b:240226816
TEST=hps --mcp cmd reset && hps --mcp dl 0 /firmware/hps/mcu_stage1.bin

Change-Id: I8e31e4b079ea1bea1c6176525d50fab713afc627
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3784521
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>