multiple: Support multi-policy signature sealing for TPM1.2
We don't need to stick with the default & extended policy since we have
the flatbuffer generator to generate to serialization code.

This is the last change to change the format of Tpm2PolicySignedData, we
will not be able to change the format after USS launch.

BUG=b:229523614
TEST=tast run $DUT hwsec.Challenge*

Change-Id: I48b970611a8b7fc1dc25af73ea56bfa2817c4eab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3787974
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Tested-by: Yi Chou <yich@google.com>
Commit-Queue: Yi Chou <yich@google.com>