federated: replace the terminology SessionName
In ChromeOS design a client corresponds to a task group deployed on the
server side. SessionName in fcp library is not a critical argument ( it
can be empty) and cannot serve as task identifier. We can achieve this
by passing in the client name as population name (and registering only
one task under one population on the server side).

This CL doesn't change much except for some names. The real trick
happens in the fcp library which now treats the string as population
name instead of session name.

BUG=b:239623649
TEST=cros_workon_make --board=hatch --test federated-service
TEST=Verified on DUT against the local brella server, too

Change-Id: I12e3304731cdfa7cef376ff5c49b66d5d0706e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3787975
Reviewed-by: Dmitry Grebenyuk <dgrebenyuk@google.com>
Tested-by: Xinglong Luan <alanlxl@chromium.org>
Commit-Queue: Xinglong Luan <alanlxl@chromium.org>