hps: avoid unnecessary power-on delay
There is no reason to sleep after powering on the HPS peripheral. The
CheckMagic() function already waits up to 3 seconds for the peripheral
to start answering on I2C.

When the HPS firmware is already up-to-date, this change reduces the
total boot time from 7.1s to 6.1s. Tested against both stage0 v4 and v3.

BUG=b:240226816
TEST=cros deploy $DUT hpsd && ssh $DUT restart hpsd

Change-Id: I3f742f66bcbf6a95d90cd7ffafe0ee4e055ddeea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3784524
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>