trunks: treat VERSION_MISMATCH as successful ping too
The ping action is to determine server PinWeaver version, so that we can
use the minimum of server and client's version for upcoming
communication. However, currently trunks return error when pinging if
server version is lower than us, and we'll force fallback to version 0.
When the default version becomes larger than 1 in the future this will
introduce a bug.

BUG=b:240376821
TEST=trunks unittest
TEST=change default version to 100 in libhwsec, then see it falling back
to v1, not v0, in runtime.

Change-Id: I3063d2c9423fc5ce3ad5f0afaedf81017a08e4df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3790322
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>