runtime_probe: Use minigbm_helper to check dgpu
Use minigbm_helper to check the drm device to see if it is a dGPU or a
iGPU. It is done by checking the vendor / driver and others. The logic
will be kept in minigbm_helper.
Also probe pci subclass 0x02 for 3D controller devices.

BUG=b:240390351
TEST=FEATURES=test emerge-zork runtime_probe
TEST=Try to probe gpu on DUT.

Cq-Depend: chromium:3784502, chromium:3780065
Change-Id: Id3242489e6ad136084b2fb2c5c067e517051df11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3789302
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>
Tested-by: Chung-Sheng Wu <chungsheng@google.com>