sepolicy: Add cdm-oemcrypto-wv14 variant
We are temporarily adding a clone of the cdm-oemcrypto binary that is
called cdm-oemcrypto-wv14 which will support the WV14 FW on
volteer/brya. The default config will support WV16. We are just adding
the binary here to group it in with the same policy as the existing
binary.

BUG=b:240453750
TEST=CQ

Cq-Depend: chrome-internal:4900345
Change-Id: Ibf875d1124d163cb171550f23b765caeb1c36760
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3789077
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>