hiberman: Restart snapshot merges upon interruption
In an earlier commit, we converted the hiberman cookie from a simple
boolean into an enum with several possible values. Let's use those
values now to recover from an unexpected reboot during the abort
process.

If we get to resume-init and find that the cookie indicates we were in
the middle of a resume or in the middle of an abort, wire up the
snapshots again so that we can merge any data in them back to the main
disk. There are a few cases to consider:

 * This is the reboot after the first attempt at a genuine resume. In
   this case, the snapshot probably has crash logs that we want, so
   reattach and sync.
 * The is a reboot after a merge was already in progress. In this case,
   the merge is likely partially completed, so we want to re-wire it up
   and continue the merge to present a consistent view of the disk (at
   least before the crash).
 * This is not the first unexpected reboot that occurred while we've
   been trying to merge the snapshots. Here we have to get our crystal
   ball out. If the unexpected reboots are unrelated to the snapshot
   merge itself, then the right thing to do is wire up the snapshots
   again and continue trying to merge. However if attempting the merge
   is in fact causing the crash, we are damned if we merge and probably
   damned if we don't. Maybe self repair is the right answer at that
   point, though it's hard to have confidence that the merge itself is
   causing the crash.

Despite some corner cases to work out, this at least wires up the common
path of retrying a merge across an unexpected reboot during resume.

BUG=b:233103309
TEST=Merge and surprise reboot.

Change-Id: I22f1cf4ee9970e9312b1ca8396a779f9a4d0fef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3794297
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>