cryptohome: Implement ListAuthFactors API for configured_auth_factors
Adds the implementation of the ListAuthFactors operation. The operation
is pretty basic: take the input user, construct auth factor objects for
all of the factors we can find for them, and use that to fill in the
response.

This does not include an implementation of the supported_auth_factors
field which will come separately.

BUG=b:208357938
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I883d5964a6b189ec815299413819cd33f8c49262
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3794303
Auto-Submit: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>