shill: portal: unique Start/StopPortalDetection method
This patch changes Device::StartPortalDetection to also stop portal
detection if the Service configuration state or Manager configuration
state has changed and Service::IsPortalDetectionDisabled() returns true.
In that case the connection state of the Service is set to "online". The
name of this method is changed to UpdatePortalDetector() to reflect
the new behavior.

The order of checks in UpdatePortalDetector() is now the following:
  - check that the Service is connected.
  - check for any conditions where portal detection should stop or not
    be running
  - check if portal detection should not be restarted.
  - start or restart portal detection.

This change allows in follow-up patches to unify all triggers starting or
stopping portal detection to UpdatePortalDetector() with no extra state
introspection.

BUG=b:230339227
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I3df3c861d02949da0d9bacdb05273a63f9e1e147
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3802209
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>