shill: define Device::OnNetworkValidationStop event
This patch adds a OnNetworkValidationStop() event on the base Device
class the subclasses can listen to to be notified when PortalDetector is
stopped.

If portal detection stops because it successfully verified Internet
access, OnNetworkValidationSuccess() is called first immediately before
OnNetworkValidationStop().

If portal detection fails to verify Internet access, reschedule an
attempt but fails to restart (example: incorrect newly selected probe
URL) and therefore stops, OnNetworkValidationFailure() is called first
immediately before OnNetworkValidationStop().

If portal detection fails its first attempt when starting or restarting
and therefore stops, OnNetworkValidationFailure() is not called because
OnNetworkValidationStart() is also not called.

Finally, when portal detection is stopped and the connection state of
the Service changes, OnNetworkValidationStop() is called before
Service::SetState().

BUG=b:216351118
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I60a9621a4b4bc1287ac21f96c9ed04e8d8a042e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3802210
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>