hiberman: Emit hibernate-tpm-done event on resume
Hiberman, or rather the kernel, uses the TPM early in the resume
process when kernel-hardening is enabled to decrypt the sealed hibernate
key. This races with trunksd, which squats on all the available TPM
handles if it gets fired up first. When USE=hibernate is set, upstart
will wait to start trunksd until hiberman resume is done going through
the TPM unseal operation.

Add that code in hiberman to emit the event upstart is waiting for to
start trunksd. This event is emitted both in the "live" resume path
after using the TPM, and in the case where there is no image to resume
to. This is done because hiberman still sticks around through login to
collect the secret seed from cryptohome, and we wouldn't want to block
trunksd all this time.

Note that trunksd also accepts "stopped hiberman" as a substitute event
for the hibernate-tpm-done event. So it's not critical that this event
get emitted on every error path (hence not using Drop/RAII here), just
any path where we'll stay alive and block for a long time.

BUG=b:237086416
TEST=Reboot with no resume pending, and resume from hibernate

Change-Id: I0749364740cff380a1c0a33dc0e488458ee850ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3803806
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>