trunks: Don't parse HE secret in reset_auth
We will be removing this field from the response after protocol version
2, and though backward compatibility is preserved in the structure, the
returned HE secret is zeroed. Prior to that, we should eliminate the
dependency of the returned HEC.

Currently the exposed APIs in libhwsec already don't take HEC in
ResetCredential, so it's mainly just trunks interface change, and a
minor modification to pinweaver_client test util.

BUG=b:241059840
TEST=hwsec.PINWeaver
TEST=pinweaver_client selftest
TEST=trunks & libhwsec unittest

Change-Id: Ib42daad3f18e2f5a37b760e21211d0eb2317b156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3802762
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>