camera: dup buffer FDs for Mojo transport
Mojo closes the FDs in a mojo callback, so we need to dup the FDs if
we want to keep the original FDs open.

While we're at it, copy the checks of close() return value from
base::ScopedFD to camera_buffer_handle so that we can catch
double-close bugs more easily in the future.

BUG=b:241459028
TEST=Run cros_camera_connector_test on redrix

Change-Id: I9f97d647cd776dd757ba726b7ac5ae13f2198b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3811049
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>