hiberman: Update UswsuspUserKey to match kernel v2
Upstream (rightfully) suggested making the key array in uswsusp_user_key
32 bytes instead of 16 bytes. The kernel currently only uses 16 of those
bytes due to the hardcoded choice of gcm(aes) with AES128, but that's an
implementation detail, rather than something to codify in an interface
between the kernel and userspace.

Update the hiberman version of this structure to match the v2 version
going out on the kernel list at some point. The source for this key is
our metadata structure, which currently supports both user-based
encryption and kernel-based encryption. Rather than chasing the 16->32
byte conversion all throughout the metadata and its users, allow
UswsuspUserKey to initialize using the length the user supplied, which
may be less than 32.

BUG=b:237086416
TEST=Hibernate and resume with kernel v2 series

Change-Id: Iceb6564384d3c5b3f44132d5ac6428efe6e07ae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3835336
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>