hiberman: Timeout and resiliency during snapshot merging
Add a timeout to the wait for snapshot merge. Because this is
effectively waiting on disk I/O, I wanted to choose a silly high number
such that even if the disk were on fire it could still sync nearly full
snapshots in this time. I chose 20 minutes, which assuming a totally
full snapshot set of 2GB means the minimum acceptable rate is ~1.7MB/s.

Additionally, rework the order in which certain operations happen on
abort. I ran into an error where the abort started, merge completed, but
the cookie didn't get cleared, which was a bug. But it also made me
realize the startup script doesn't go well unless resume-init manages to
wire up all 3 snapshots correctly (if it indicates it wired up any).

So to make this more robust, make sure that we don't delete any
snapshots until we've merged them all and cleared the hibernate cookie.
So the order of operations is now:
 * Kick off merges of all snapshots
 * Record any that failed to kick off
 * Wait for merges to complete, or time out
 * Change the hibernate cookie to NoResume so snapshots don't get wired
   up again.
 * Delete all snapshots

BUG=b:241432906
TEST=Hibernate and abort in resume

Change-Id: I04d8afb1c9b885b3f17e68e7686973cb9d1caa92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3835337
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>