installer: Avoid failing UpdateLegacyKernel during chromeos-install.
After booting from a USB mass storage device and attempting to
chromeos-install to internal storage, EfiPostInstall fails.
EfiPostInstall fails because UpdateLegacyKernel is attempting to copy
vmlinuz from the /boot directory into the newly created EFI boot
partition. However, vmlinuz.A and vmlinuz.B have already been created
in the appropriate location earlier in the installation process.

Change the behavior to return true if the destination exists, if the
source does not. This handles the case where there is no new Kernel
to update. Said another way, when the Kernel on the EFI boot partition
is the new Kernel (as is the case during install).

BUG=b:241554104
TEST=build_packages && build_image && cros flash usb://
     Boot via USB on an EFI based platform.
     chromeos-install --dst /dev/nvme0n1

Change-Id: I036b7f0e50232137ad02be6bfd3bb98d89e53617
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3813888
Tested-by: Eric Rathja <eric.rathja@intel.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Kevin Derossett <kevin.l.derossett@intel.corp-partner.google.com>