shill: Use OnceCallback in the done callback of wake on wifi
The `done_callback` is called and reset so it can be converted to an
OnceCallback. Also rename the helper function from `RunAndResetXXX` to
`RunXXX` since the reset is no longer needed.

This is a part of the base::Callback cleanup project.
See go/chromeos-code-health-callback for more details.

BUG=chromium:1205291
TEST=USE="-cros-debug" FEATURES=test emerge-volteer shill

Change-Id: I0d00eeee31a500bc42c7ec443495602316bb3b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3815634
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>