arc-adbd: Fix missing seccomp file
The seccomp policy file for arc-adbd in ARC-P has been
missing due to a config change in crrev.com/c/3403461.
This causes the service to crash as soon as it
starts.

This CL fixes that to put up the seccomp file at the
desired location.

BUG=b:241957917
TEST=Manual. Service does not crash due to seccomp now.

Change-Id: I6d8bad33cc2b5ecd38398a05f9bf4d0d2f286c1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3834715
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Prameet Shah <phshah@chromium.org>