shill: Track supported security ciphers in wifi.cc
Add a private variable to WiFi device object to record the support for
security ciphers. This support is read in from the NL80211 attribute
NL80211_ATTR_CIPHER_SUITES, which is included in NL80211_CMD_NEW_WIPHY.

WEP support is the impetus for this change, and this change also adds a
function to report if WEP is supported.

Design doc: go/cros-platform-wep-deprecation-dd

BUG=b:233243778
TEST=FEATURES=test emerge-hatch shill
TEST=Deploy to hatch device; shows WEP support.
TEST=Deploy to nipperkin device; shows no WEP support.

Change-Id: I0c9697190f057121ef69c461d5a9179834d3009c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3821086
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: David Ruth <druth@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>