login: Add support for explicit migration mode
We are introducing a new migration mode for Chrome. Instead of
using a is_move argument, move to a more generic mode argument.
Since this is an incompatible change, keep parsing the DBus message
manually in StartBrowserDataMigration. The workaround will be removed
after Chrome is updated.

Changes to Chrome are made in crrev.com/c/3820120.

BUG=b:242003477
TEST=TEST=cros_run_unit_tests --board $BOARD --packages chromeos-login,
TEST=tast run $DUT lacros.Migrate.* with and without crrev.com/c/3820120

Change-Id: I9bed1e4eb1cb3e63682962e7836aa458344f038a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3821862
Reviewed-by: Yuta Hijikata <ythjkt@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Victor-Gabriel Savu <vsavu@google.com>