vm_tools: sommelier: Pipe surface to transform logic
The surface on which transformations are being performed on
will be passed into the sl_transform_* functions.

Currently nothing is being done with them and this will be
addressed within a subsequent patch in this series.

This is so that we can override the global XDG scaling factors
with (slightly modified) ones that are surface specific.
This is done in cases where the window width/heights cannot be
exactly represented in the logical coordinate space.

BUG=b:217976094
TEST=Verified that Sommelier functioned in direct/non-direct mode.

Change-Id: Ia3cac59f758c14a85df096257e7f971a2c01be59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3833448
Commit-Queue: Isaac Bosompem <mrisaacb@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Tested-by: Isaac Bosompem <mrisaacb@google.com>