shill: cellular: Fix MTU to be used in FirstNet
FirstNet requires an MTU of 1342, which is usually negotiated
automatically with the network.

The Fibocom L850 does not correctly report the network MTU in certain
old firmware versions, so it is required to have the value hardcoded
for the FirstNet operator in the mobile operator database.

BUG=b:242087091
TEST=None

Change-Id: I0875f1509806f64f16a841997c40754ca2c20962
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3827343
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Auto-Submit: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>