crash: Better signature for Rust crashes
This CL builds on crrev.com/c/3834430 to improve crash signatures for
rust apps. The custom panic handler in that CL exposes a crash signature
via a memfd. The user collector uses that signaure if it is found.

For now, this is not automatic. Rust applications will have to call the
libchromeos::panic_handler::install_memfd_handler() function at startup
to get this behavior.

BUG=b:234093439
TEST=Added unit test. See crrev.com/c/3836804 for the integration test.

Change-Id: I65f041a7560fe672e45078bfe92892410373cfc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3834930
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Paramjit Oberoi <psoberoi@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Paramjit Oberoi <psoberoi@google.com>