vm_tools: cicerone extract vm information for shader cache install
This is the implementation for go/borealis-shader-client-design.

container_listener delegates rest of the actions to cicerone service.

In later CLs, cicerone service will communicate with concierge to get
gpu cache path (for mounting), which then will be passed along to the
daemon responsible for DLC installations and mounting.

BUG=b:239494222
TEST=In borealis, /opt/google/cros-containers/bin/garcon --client
--borealis-shader-cache --app-id=620 --install --mount

Change-Id: Ia515acfc6a54f0866268142690757ce10f407c69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3823557
Commit-Queue: Max Lee <endlesspring@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Max Lee <endlesspring@google.com>