shill: Move Device::UpdateBlackholeUserTraffic() to Network (part 1)
This goal of this series of patches is to move blackhole user traffic
logic from Device to Network.

As the first step, this patch moves the logic to calculate user traffic
uids from Manager to RoutingTable. This part of code is only related to
routing and needs to be put in a singleton for the caching purposes, and
thus RoutingTable is a good place to hold it.

No behavior change will be introduced by this patch.

BUG=b:232177767
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I66f8d75797f913ff1d127875d83f5ba8b72c8a0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3826860
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>