secagentd: initial checkin
check in the basics for secagent

* bpf skeleton generation tools
* a simple main
* a bare bones BUILD

The gni for bpf skeleton and the python skeleton generator will
eventually be migrated to the common area once it is stable.

This is to avoid other packages forming dependencies on a tool-chain
that is still in flux.

BUG=b:240344505
TEST=generated some BPF programs and they work fine

Cq-Depend: chromium:3831431
Change-Id: I4896dfff1b8d6c14f1cf361dd3d1030851a761f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3833205
Commit-Queue: Jason Ling <jasonling@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Tested-by: Jason Ling <jasonling@chromium.org>