hiberman: Don't open metrics on non-resume boots
Hiberman is exiting prematurely on systems where no hibernation has ever
been attempted. This is because it tries to open a metrics file in the
hibernate directory, which has not yet been created (as it gets created
upon first hibernation).

Move the metrics setup down after the critical resume functionality
(clearing the hibernate cookie) has finished, and a resume is actually
going to happen. This is better for resilience and probably makes boot
just a tiny bit faster in non-resume cases.

BUG=b:232184089
TEST=boot on a system with no hibernation, 'status hiberman' running.

Change-Id: Ibb526732a4af756a05e78150f5e813efed08642e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3835335
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>