cryptohome: Add metrics for replay operations
Add metrics in ReplayLogEntries to record the results of replay
operations, as it's important information to the stability of PinWeaver.
Use a separate enum for the result because there are many possibilities
we want to individually track, and it's not suitable to abuse the
LECredError enum. Also, separate the histograms between a normal replay
and a full replay that uses up all the entries because in the latter,
it's just a best-effort attempt hoping that we're only 1 entry behind
the first log entry, and the error distribution might be very different.

Separate metrics for individual replay operations from the Sync
operation, to monitor them better. A special note is that we separate
the BackendReplayLog actions for normal replay and full replay too, as
this action is affected by whether the current hash value is correct.

BUG=b:243090826
TEST=cryptohome unittests

Change-Id: Idb485761261653a1dad76acdd139029fd402f11c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3842250
Reviewed-by: John L Chen <zuan@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>