shill: Do not trigger OnNetworkStopped() in Network::Start()
In the current implementation of Network, Stop() is called at first in
Start(), and this will trigger the OnNetworkStopped() callback which
should be unexpected. This patch fixes this behavior.

Note that the current behavior does not indicate any bugs now since
Device::OnNetworkStopped() is not doing anything unexpected.

Also add unit tests for Network to cover the added logic in this patch.

BUG=b:232177767
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: I527236903b2041fbb6844b07d1700832e9241de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3842260
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>