shill: portal: accept 307 response for captive portals
Some portals use a 307 temp redirect http response to bring users to the
signin page. This updates the portal detector to accept a 307 similarly
to a 302, making service state `redirect-found`. It also checks for the
redirect url to be valid. If invalid, the state will be
`portal-suspected`.

BUG=b:172216846
TEST=FEATURES=test P2_TEST_FILTER=*Portal* emerge-${BOARD} shill, also
added two new tast tests to ShillCaptivePortalHTTP.go to verify that 307
will have a `redirect-found` state with a valid url and a
`portal-suspected` with an invalid url (https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3842598).

Change-Id: I346ed51c63d19bf5fcb2a0858da7e76375855636
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3840985
Tested-by: Michael Rygiel <michaelrygiel@google.com>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>