cryptohome: Add support for application container reset
With the introduction of application containers for
cryptohome vaults, we need a mechanism for resetting the
storage of the container without needing to re-instantiate
the encrypted container itself. This allows applications to
reset their storage in the middle of a session without
needing access to the user's file encryption keys.

This CL adds Reset to encrypted containers and exposes this
via CryptohomeVault. A subsequent CL wires this up with
the cryptohome daemon via a D-Bus method.

BUG=b:241302753
TEST=unittests

Change-Id: I50af83901426b326686d4d900ad8eb3fba2b5810
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3843175
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>