chromeos-config: Add support for FRID matching
On ARM devices, we currently implement matching based on device-tree
compatible string.  This has worked for the most part so far, but we
don't always control the list of compatible strings as we wish.  See
the bug for a good example case where compatible is insufficient.

There have been previous discussions of switching to FRID matching for
both x86 and ARM in 2019, as well as an implementation in mosys that
never was fully flushed out.

This re-implements the idea in the crosid world, for both x86 and ARM.
I also intend to deprecate SMBIOS name and fdt-compatile matching in
favor of FRID matching in all configs in the future.

BUG=b:242681769
TEST=provided unit tests pass

Change-Id: I6bf952726cac119c518e05161ce8a6add113a11a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3843176
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>