shill: strip whitespace from untrusted attribute
Strip the trailing '\n' in the pci sysfs `untrusted`
attribute file.
This was causing shill to attempt to disable offloading on
an internal (untrusted == 0) pci device.
The stmmac driver always fails disabling these features
prior to bringing up the interface so the internal ethernet
device would never be brought up.
Fix the check so it no longer attempts to disable offloading
for this internal device allowing it to be used.

BUG=b:243203825,b:228327681
TEST=FEATURES=test emerge-reven shill
TEST=Manually confirm eth works on a reven stmmac ethernet device

Change-Id: I28fbc0d91597162ca7e17331f5b7788ebb52c0f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3843388
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>