shill: Increase LOG level about addresses in DeviceInfo
Increase the level of several LOGs in DeviceInfo about adding or
removing addresses. They should only show up when there is an IP address
change in the kernel or shill tries to change the address, so will not
be very noise. This should be helpful when debugging races between the
address cache in shill and the kernel.

BUG=b:239753191
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: I18219179779d3e276e5385fb25a07a40136e43b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3844736
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>