hiberman: Add demand start dbus service
Chrome calls the ResumeFromHibernate() dbus call whenever a user
successfully authenticates. This is needed not only at the boot login to
initiate a resume from hibernate, but also after subsequent logins to
set up the following hibernation. Today, on a hibernate-enabled system,
a user that logs out and tries to log in again will get stuck at the
login prompt, as Chrome is stuck waiting for a service that's ended.

Demand start hiberman upon subsequent logins so that newly logged in
user can hibernate.

BUG=b:240142469
TEST=log out and log in on a brya-hibernate image

Change-Id: Id84c5883017c7c2ab478c6ecdfbd49305f16b013
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3846424
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>