system_api: Proto definition for AuthIntent
Add proto definition of the AuthIntent enum and its usages into
cryptohome's AuthSession API.

The intent is a declaration of which operations the caller (Chrome)
wants to perform after the authentication succeeds. Specifying it in
advance, when starting the AuthSession, allows Cryptohome to choose
eligible factors. For example, an unlock-only factor (like a legacy
fingerprint or ephemeral user's password) can be used for Lock Screen or
for visiting the web Password Manager, but not for the login.
Additionally, it allows cryptohome to choose a faster cryptographic
scheme if the intent allows so (e.g., the VERIFY_ONLY intent allows a
quick in-memory credential check).

We start from the minimal set of intents to support main scenarios
(unlocking using unlock-only factors): DECRYPT and VERIFY_ONLY. In the
future, more fine-grained intents can be added.

BUG=b:240596931
TEST=emerge-$BOARD chromeos-base/system_api

Change-Id: I86d7c2d86436e67ebfdbba75f52992e9bb25de33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3852836
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>