lorgnette: Block MF260 from pixma backend
The MF 260 and MF 260 II series both work with eSCL but not the pixma
backend.  Even though eSCL is supposed to be preferred, there's no
benefit to including the pixma responses in the scanner list at all.

Rather than just hardcoding the MF260 case, this adds a slightly more
general function that can be used to exclude other devices in the
future if needed.

BUG=b:233012341
TEST=FEATURES=test emerge-asurada lorgnette

Change-Id: I146e012c73f49ea88447448a656899036952b390
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3854552
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>