hiberman: Factor out snapshot name enumeration
We already had three places where we've copy/pasted a few lines of code
to enumerate the snapshot directory and iterate over the file names. I'm
about to add a fourth, so factor out those lines into a function that
returns a vector of the file name strings in the snapshot directory.

BUG=b:237011312
TEST=cargo build

Change-Id: I095c54fca50f483e3802587f7e17874aeaef03cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3859261
Reviewed-by: Alexis Savery <asavery@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>