cryptohome: Populate the reset seed when constructing a PIN vault keyset
Currently, AddAuthFactor does not work with PINs, due to a lack of a
reset seed. I'm not sure if there's a better way to do this, but at
least right now given that we already have a special conditional for the
PIN, we can insert some logic there.

BUG=b:240931575
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I7943afd50b6a1883699d181fb35061e58d43755c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3855019
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>