cros-disks: Try to unmount with default options first
When unmounting a mount point, try to unmount it in a gentle way with
default options first. If this fails with the EBUSY error, then try to
forcefully unmount it with the MNT_FORCE and MNT_DETACH options.

There is a suspicion that unmounting with the default options might
prevent some filesystem corruptions. This is not proven yet, but as a
precaution, we take this extra step before forcefully unmounting.

This CL reverts the effects of CL 3631745.

BUG=None
TEST=tast run dut platform.CrosDisks*

Change-Id: Ief59ecadbb8e3780d15175dabfba7f142948228c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3855821
Reviewed-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>