init: Add simple OS caching layer to boot-update-firmware
Most of the ChromeOS firmware updater scripts are run from here, and
are blocking the UI. The CSME firmware updater in particular may take
more time to verify whether or not it needs to update the CSME
firmware, therefore this patch adds some logic to cache the OS version
of the current boot. Upon boot, if this cached version is different
than the running OS, then the CSME firmware updater script is run; if
the versions match, then it is skipped under the assumption that the
CSME firmware cannot be updated without a resulting OS version bump
too. The cached version is updated only if the CSME firmware updater
was successful.

Cq-Depend: chromium:3929101
BUG=b:238908603
TEST=Install onto test system (agah), reboot twice, and see
the following in /var/log/messages:
  Update FPMCU firmware.
  Update CSME firmware.
  Update fwupd firmware.
  Update FPMCU firmware.
  CSME firmware updated already for this OS version; skipping until next OS update.
  Update fwupd firmware.
TEST=Observe that /var/lib/boot-update-firmware/last-os-version has the
correct OS version in it after a successful run.
TEST=Run `echo '' > /var/lib/boot-update-firmware/last-os-version`, and
see the same sequence of prints after the next two reboots.
TEST=Inject an error into the CSME update script and observe that
/var/lib/boot-update-firmware/last-os-version is not updated.

Change-Id: I39e34d82c9080e10132d024bf3e736dc54e40c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3859311
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Tarun Tuli <taruntuli@google.com>
Tested-by: Tarun Tuli <taruntuli@google.com>